+++ solver_test.c
@@ -4,14 +4,14 @@
 #include <unistd.h>
 #include "mmio.h"
 
-static void compare(double* yResult, double* yReorder, const double threshold, const int dimension)
+static void compare(float* yResult, float* yReorder, const float threshold, const int dimension)
 {
-	double avgdiff = 0; 
-	double avgampldiff = 0;
+	float avgdiff = 0; 
+	float avgampldiff = 0;
 	for (int i = 0; i < dimension; ++i)
 	{
-		double d = fabs(yReorder[i] - yResult[i]);
-		double ampl = fmin(fabs(yReorder[i]), fabs(yResult[i]));
+		float d = fabs(yReorder[i] - yResult[i]);
+		float ampl = fmin(fabs(yReorder[i]), fabs(yResult[i]));
 		if (d > ampl*threshold)
 		{
 			printf("large difference at %d  : %f vs %f\n", i, yReorder[i] , yResult[i]);
@@ -24,18 +24,18 @@ static void compare(double* yResult, double* yReorder, const double threshold, c
 	printf("diff is %e, ampldiff is %e\n", avgdiff, avgampldiff);
 }
 
-static int matrixRead_unsym(matrixCOO* localMatrixCOO, double** xCompare_in, double** y_in, FILE *f)
+static int matrixRead_unsym(matrixCOO* localMatrixCOO, float** xCompare_in, float** y_in, FILE *f)
 { 	int _dimension, _N;
 	int ret_code, totalNum;
 	if ((ret_code = mm_read_mtx_crd_size(f, &_dimension, &_N, &totalNum)) !=0)
 		exit(1);	
 	localMatrixCOO->dimension = _dimension;
-	*xCompare_in = (double*)malloc(_dimension*sizeof(double));
-	*y_in = (double*)malloc(_dimension*sizeof(double));
-	double*	xCompare = *xCompare_in;
-	double*	y = *y_in;
+	*xCompare_in = (float*)malloc(_dimension*sizeof(float));
+	*y_in = (float*)malloc(_dimension*sizeof(float));
+	float*	xCompare = *xCompare_in;
+	float*	y = *y_in;
 	localMatrixCOO->totalNum = totalNum;
-	localMatrixCOO->nParts = ceil(((double) _dimension)/vectorCacheSize);
+	localMatrixCOO->nParts = ceil(((float) _dimension)/vectorCacheSize);
 	printf("parts is %d\n", localMatrixCOO->nParts);
 	localMatrixCOO->partBoundary = (int* )calloc(_dimension, sizeof(int));
 	localMatrixCOO->reorderList = (int* )calloc(_dimension, sizeof(int));
@@ -44,23 +44,23 @@ static int matrixRead_unsym(matrixCOO* localMatrixCOO, double** xCompare_in, dou
 
 	localMatrixCOO->I=(int *) malloc(totalNum*sizeof(int));
 	localMatrixCOO->J=(int *) malloc(totalNum*sizeof(int));
-	localMatrixCOO->V=(double *) malloc(totalNum*sizeof(double));
-	localMatrixCOO->diag = (double *) malloc(_dimension*sizeof(double));
+	localMatrixCOO->V=(float *) malloc(totalNum*sizeof(float));
+	localMatrixCOO->diag = (float *) malloc(_dimension*sizeof(float));
 
 	int* numInRow = localMatrixCOO->numInRow;
 	int* I = localMatrixCOO->I;
 	int* J = localMatrixCOO->J;
-	double* V = localMatrixCOO->V;
+	float* V = localMatrixCOO->V;
 
 	for (int i=0;i < _dimension;i++){
 		srand(i);
-		xCompare[i]=(double) (rand()%200-100)/100;
+		xCompare[i]=(float) (rand()%200-100)/100;
 		//x_compare[i]=1;
 	}
 	int tempI, tempJ;
-	double tempV;
+	float tempV;
 	for (int i=0; i<totalNum; i++){
-		fscanf(f, "%d %d %lg\n", &tempI, &tempJ, &tempV);
+		fscanf(f, "%d %d %f\n", &tempI, &tempJ, &tempV);
 		J[i]=tempJ-1;  /* adjust from 1-based to 0-based */
 		I[i]=tempI-1;
 		V[i]=tempV;
@@ -84,7 +84,7 @@ static int matrixRead_unsym(matrixCOO* localMatrixCOO, double** xCompare_in, dou
 	localMatrixCOO->maxCol = maxCol;
 	return 1;
 }
-static int matrixRead_sym(matrixCOO* localMatrixCOO, double** xCompare_in, double** y_in, FILE *f)
+static int matrixRead_sym(matrixCOO* localMatrixCOO, float** xCompare_in, float** y_in, FILE *f)
 {
 	int _dimension, _N, _lowerNum;
 	int ret_code, totalNum, lowerNum;
@@ -94,10 +94,10 @@ static int matrixRead_sym(matrixCOO* localMatrixCOO, double** xCompare_in, doubl
 	lowerNum = _lowerNum;
 	totalNum = lowerNum*2-_dimension;
 	
-	*xCompare_in = (double*)malloc(_dimension*sizeof(double));
-	*y_in = (double*)malloc(_dimension*sizeof(double));
-	double*	xCompare = *xCompare_in;
-	double*	y = *y_in;
+	*xCompare_in = (float*)malloc(_dimension*sizeof(float));
+	*y_in = (float*)malloc(_dimension*sizeof(float));
+	float*	xCompare = *xCompare_in;
+	float*	y = *y_in;
 	/*The overall number of nozeros in this matrix*/
 	localMatrixCOO->totalNum = totalNum;
 
@@ -110,22 +110,22 @@ static int matrixRead_sym(matrixCOO* localMatrixCOO, double** xCompare_in, doubl
 
 	int* lowerI=(int *) malloc(lowerNum*sizeof(int));
 	int* lowerJ=(int *) malloc(lowerNum*sizeof(int));
-	double* lowerV=(double *) malloc(lowerNum*sizeof(double));
+	float* lowerV=(float *) malloc(lowerNum*sizeof(float));
 
 	localMatrixCOO->I=(int *) malloc(totalNum*sizeof(int));
 	localMatrixCOO->J=(int *) malloc(totalNum*sizeof(int));
-	localMatrixCOO->V=(double *) malloc(totalNum*sizeof(double));
-	localMatrixCOO->diag = (double *) malloc(_dimension*sizeof(double));
+	localMatrixCOO->V=(float *) malloc(totalNum*sizeof(float));
+	localMatrixCOO->diag = (float *) malloc(_dimension*sizeof(float));
 	int* numInRow = localMatrixCOO->numInRow;
 	int* I = localMatrixCOO->I;
 	int* J = localMatrixCOO->J;
-	double* V = localMatrixCOO->V;
-	double* diag = localMatrixCOO->diag;
+	float* V = localMatrixCOO->V;
+	float* diag = localMatrixCOO->diag;
 	
 	int tempI, tempJ;
-	double tempV;
+	float tempV;
 	for (int i=0; i<lowerNum; i++){
-		fscanf(f, "%d %d %lg\n", &tempI, &tempJ, &tempV);
+		fscanf(f, "%d %d %f\n", &tempI, &tempJ, &tempV);
 		lowerJ[i]=tempJ-1;  /* adjust from 1-based to 0-based */
 		lowerI[i]=tempI-1;
 		lowerV[i]=tempV;
@@ -155,7 +155,7 @@ static int matrixRead_sym(matrixCOO* localMatrixCOO, double** xCompare_in, doubl
 	numInRow[_dimension-1]=0;
 	for (int i=0;i < _dimension;i++){
 		srand(i);
-		xCompare[i]=(double) (rand()%200-100)/100;
+		xCompare[i]=(float) (rand()%200-100)/100;
 		//x_compare[i]=1;
 	}
 	int index1, index2;
@@ -197,8 +197,8 @@ int main(int argc, char* argv[])
 	MM_typecode matcode;
 	int MAXIter = 0;
 	FILE *f;
-	double *y;
-	double *xCompare;
+	float *y;
+	float *xCompare;
 	char fileName[100];
 	fileName[0] = '\0';
 	int oc;
@@ -281,7 +281,7 @@ int main(int argc, char* argv[])
 		matrixRead_unsym(&localMatrixCOO, &xCompare, &y, f);
 	}
 	fclose(f);
-	double *yResult = (double *) calloc(localMatrixCOO.dimension, sizeof(double));
+	float *yResult = (float *) calloc(localMatrixCOO.dimension, sizeof(float));
 
 	//spmvHYB(&localMatrixCOO, y, x, MAXIter);
 	//solverGPuUnprecondCUSPARSE(&localMatrixCOO, y, x, MAXIter);
@@ -291,8 +291,8 @@ int main(int argc, char* argv[])
 	//}
 	//return 0;
 
-	double *xReorder = (double* )calloc(localMatrixCOO.dimension, sizeof(double)); 
-	double *yReorder = (double* )calloc(localMatrixCOO.dimension, sizeof(double)); 
+	float *xReorder = (float* )calloc(localMatrixCOO.dimension, sizeof(float)); 
+	float *yReorder = (float* )calloc(localMatrixCOO.dimension, sizeof(float)); 
 	if (mm_is_symmetric(matcode)){
 		matrixReorder(&localMatrixCOO);
 	} else {
@@ -330,7 +330,7 @@ int main(int argc, char* argv[])
 	//interval2=(end_time2-start_time2)*1000/CLOCKS_PER_SEC;
 
 	//printf("time consuming CPU is %f, time consuming GPU is %f, speedup is %f\n", interval1, interval2, interval1/interval2);
-	//double Gflop=(totalNum*4+12*dimension)/interval1*1000*MAXIter;
+	//float Gflop=(totalNum*4+12*dimension)/interval1*1000*MAXIter;
 	//printf("error is %f, total num is %d, time is %f ms, Gflops is %f, final error is %f\n",result_error/dimension, totalNum, interval1, Gflop, error_track[MAXIter-1]*1000);
 	return 0;
 }
